backend/fastapi-service/requirements.txt
fastapi
uvicorn
pydantic
requests
groq
python-dotenv
pandas
